requirements.txt
langchain-core
langchain-community
langchain[mistralai]
wikipedia
python-dotenv
pydantic
duckduckgo-search
streamlit